[sanjeeva] : checked 3 test cases by giving max values at 3 positons

+++ MaxOfThree.cs
@@ -8,7 +8,7 @@ namespace GenericsAssignment
 {
     internal class MaxOfThree
     {
-        public static int get_max(int a,int b,int c)
+        public static float get_max(float a,float b,float c)
         {
             if (a.CompareTo(b) > 0 && a.CompareTo(c) > 0 ||
                 a.CompareTo(b) >= 0 && a.CompareTo(c) > 0 ||
